Make code more DRY.
